requirements.txt
pandas
geopandas
shapely
matplotlib
pyjanitor
folium
matplotlib
mapclassify
streamlit
streamlit_folium
black
